requirements.txt
pypdf==4.0.0
PyMuPDF==1.23.8
pdf2image==1.17.0
Pillow==10.2.0
reportlab==4.0.9